--- assets/dev/js/app/app.js
@@ -1,11 +1,0 @@
-var App = Marionette.Application.extend( {
-
-	onStart: function() {
-		this.initComponents();
-	},
-
-	initComponents: function() {
-	}
-} );
-
-module.exports = ( window.elementorApp = new App() ).start();
