+++ assets/dev/js/common/common.js
@@ -1,0 +1,5 @@
+class ElementorCommonApp extends Marionette.Application {
+
+}
+
+window.elementorCommon = new ElementorCommonApp();
